Allow relay control via C-r and C-t
C-r and C-t (DC2 and DC4) will now enable and disable
the relays for the current loop, and motor, respectively.

Enabling the relays will allow a four second pause between
the current loop relay enable and the motor, so it can run
closed right away. Disabling has a similar delay between
turning off the motor and the current loop for the same reason,
but it's only three seconds.

I was hoping this change would just allow the chars sent
during the delay to be buffered, but the delay is blocking,
apparently, so I'll have to reimplement the delay differently.

Likely I'll have to check for current loop status before
enabling the motor, since if the loop is closed we can check.